tracing: move some code to context.go
Release note: None
